Don't drop user-input when a form including the widget has validation errors.
At the moment the choice field is initialized too late. It won't be set during
the first call to `extract` and thus won't be saved on the widget correctly. We
move the choice-field initialization before the super call to `update` to make
sure that the choide_field is already available in the first call to `extract`.